+++ pkg/deployment/reconcile/action.register.generated.go
@@ -153,6 +153,9 @@ var (
 	_ Action        = &actionRebalancerGenerate{}
 	_ actionFactory = newRebalancerGenerateAction
 
+	_ Action        = &actionRebuildOutSyncedShards{}
+	_ actionFactory = newRebuildOutSyncedShardsAction
+
 	_ Action        = &actionRecreateMember{}
 	_ actionFactory = newRecreateMemberAction
 
@@ -759,6 +762,18 @@ func init() {
 		// Get Action defition
 		function := newRebalancerGenerateAction
 		action := api.ActionTypeRebalancerGenerate
+
+		// Wrap action main function
+
+		// Register action
+		registerAction(action, function)
+	}
+
+	// RebuildOutSyncedShards
+	{
+		// Get Action defition
+		function := newRebuildOutSyncedShardsAction
+		action := api.ActionTypeRebuildOutSyncedShards
 
 		// Wrap action main function