+++ pkg/deployment/reconciler/context.go
@@ -98,6 +98,7 @@ type ArangoAgencyGet interface {
 	GetAgencyCache() (agencyCache.State, bool)
 	GetAgencyArangoDBCache() (agencyCache.StateDB, bool)
 	GetAgencyHealth() (agencyCache.Health, bool)
+	ShardsInSyncMap() (agencyCache.ShardsSyncStatus, bool)
 }
 
 type ArangoAgency interface {
@@ -132,6 +133,8 @@ type DeploymentDatabaseClient interface {
 	// GetDatabaseAsyncClient returns a cached client for the entire database (cluster coordinators or single server),
 	// creating one if needed. Only in AsyncMode
 	GetDatabaseAsyncClient(ctx context.Context) (driver.Client, error)
+	// GetServerAsyncClient returns an async client for a specific server.
+	GetServerAsyncClient(id string) (driver.Client, error)
 }
 
 type DeploymentMemberClient interface {