+++ pkg/util/globals/global.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -29,6 +29,12 @@ const (
 	DefaultArangoDCheckTimeout   = time.Second * 2
 	DefaultReconciliationTimeout = time.Minute
 
+	// DefaultOutSyncedShardRebuildTimeout
+	// timeout after which particular out-synced shard is considered as failed and rebuild is triggered
+	DefaultOutSyncedShardRebuildTimeout = time.Minute * 60
+	// DefaultOutSyncedShardRebuildRetryTimeout timeout after which rebuild shards retry flow is triggered
+	DefaultOutSyncedShardRebuildRetryTimeout = time.Minute * 60
+
 	DefaultKubernetesRequestBatchSize = 256
 
 	DefaultBackupConcurrentUploads = 4
@@ -36,11 +42,13 @@ const (
 
 var globalObj = &globals{
 	timeouts: &globalTimeouts{
-		requests:       NewTimeout(DefaultKubernetesTimeout),
-		arangod:        NewTimeout(DefaultArangoDTimeout),
-		arangodCheck:   NewTimeout(DefaultArangoDCheckTimeout),
-		reconciliation: NewTimeout(DefaultReconciliationTimeout),
-		agency:         NewTimeout(DefaultArangoDAgencyTimeout),
+		requests:          NewTimeout(DefaultKubernetesTimeout),
+		arangod:           NewTimeout(DefaultArangoDTimeout),
+		arangodCheck:      NewTimeout(DefaultArangoDCheckTimeout),
+		reconciliation:    NewTimeout(DefaultReconciliationTimeout),
+		agency:            NewTimeout(DefaultArangoDAgencyTimeout),
+		shardRebuild:      NewTimeout(DefaultOutSyncedShardRebuildTimeout),
+		shardRebuildRetry: NewTimeout(DefaultOutSyncedShardRebuildRetryTimeout),
 	},
 	kubernetes: &globalKubernetes{
 		requestBatchSize: NewInt64(DefaultKubernetesRequestBatchSize),
@@ -108,6 +116,8 @@ func (g *globalBackup) ConcurrentUploads() Int {
 
 type GlobalTimeouts interface {
 	Reconciliation() Timeout
+	ShardRebuild() Timeout
+	ShardRebuildRetry() Timeout
 
 	Kubernetes() Timeout
 	ArangoD() Timeout
@@ -116,7 +126,7 @@ type GlobalTimeouts interface {
 }
 
 type globalTimeouts struct {
-	requests, arangod, reconciliation, arangodCheck, agency Timeout
+	requests, arangod, reconciliation, arangodCheck, agency, shardRebuild, shardRebuildRetry Timeout
 }
 
 func (g *globalTimeouts) Agency() Timeout {
@@ -129,6 +139,14 @@ func (g *globalTimeouts) ArangoDCheck() Timeout {
 
 func (g *globalTimeouts) Reconciliation() Timeout {
 	return g.reconciliation
+}
+
+func (g *globalTimeouts) ShardRebuild() Timeout {
+	return g.shardRebuild
+}
+
+func (g *globalTimeouts) ShardRebuildRetry() Timeout {
+	return g.shardRebuildRetry
 }
 
 func (g *globalTimeouts) ArangoD() Timeout {